The changes to about and making adding menu fuction
So I added form to add menus. And gave a div to the main app. So that about doesn't clip with the footer. Test this. This has absolutely no effect on the other components display.  And spruced up the about page a bit. Could bit more sprucing and searching for actual pictures to use. But overall it's done.